requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx[socks]>=0.27.0
pyyaml>=6.0.1
apscheduler>=3.10.4
